% Generated by roxygen2 (4.0.1): do not edit by hand
\name{calc_dists_dotprods}
\alias{calc_dists_dotprods}
\title{Calculate distances and dot products between two sets of neurons}
\usage{
calc_dists_dotprods(query_neurons, target_neurons, subset, ignoreSelf = TRUE,
  ...)
}
\arguments{
\item{query_neurons}{a \code{\link[nat]{neuronlist}} to use for calculating
distances and dot products.}

\item{target_neurons}{a further \code{\link[nat]{neuronlist}} to use for
calculating distances and dot products.}

\item{subset}{a \code{\link{data.frame}} specifying which neurons in
\code{query_neurons} and \code{target_neurons} should be compared, with
columns specifying query and target neurons by name, with one row for each
pair. If unspecified, this defaults to an all-by-all comparison.}

\item{ignoreSelf}{a Boolean indicating whether to ignore comparisons of a
neuron against itself (default \code{TRUE}).}

\item{...}{extra arguments to pass to \code{\link{NeuriteBlast}}.}
}
\value{
A list, for each query neuron, of vectors containing distances and
  dot products for each target neuron.
}
\description{
Calculate distances and dot products between two sets of neurons
}

