% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fill_pairs_sparse_score_mat}
\alias{fill_pairs_sparse_score_mat}
\title{Add forwards, reverse and self scores for a pair of neurons to a sparse score matrix}
\usage{
fill_pairs_sparse_score_mat(sparse_matrix, n1, n2, dense_matrix,
  reverse = TRUE, self = TRUE, reverse_self = (reverse && self))
}
\arguments{
\item{sparse_matrix}{the sparse matrix to fill in.}

\item{n1}{the name of the query neuron.}

\item{n2}{the name of the target neuron.}

\item{dense_matrix}{the score matrix from which to extract scores.}

\item{reverse}{logical indicating that the reverse score should also be filled in (default \code{TRUE}).}

\item{self}{logical indicating that the self-score of the query should also be filled in (used for normalised scores; default \code{TRUE}).}

\item{reverse_self}{logical indicating that the self-score of the target should also be filled in (used for mean scores; default \code{TRUE}).}
}
\value{
A sparse matrix (of class \code{\link[spam]{spam}}) with the specified score entries filled.
}
\description{
Add forwards, reverse and self scores for a pair of neurons to a sparse score matrix
}

