% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nblast}
\alias{nblast}
\title{Calculate similarity score for neuron morphologies}
\usage{
nblast(query, target = getOption("nat.default.neuronlist"), smat = NULL,
  sd = 3, version = c(2, 1), normalised = FALSE, UseAlpha = FALSE)
}
\arguments{
\item{query}{the query neuron.}

\item{target}{a \code{\link[nat]{neuronlist}} to compare neuron against.
Defaults to \code{options("nat.default.neuronlist")}. See
\code{\link{nat-package}}.}

\item{smat}{the scoring matrix to use (see details)}

\item{sd}{Standard deviation to use in distance dependence of nblast v1
algorithm. Ignored when \code{version=2}.}

\item{version}{the version of the algorithm to use (the default, 2, is the
latest).}

\item{normalised}{whether to divide scores by the self-match score of the
query}

\item{UseAlpha}{whether to consider local directions in the similarity
calculation (default: FALSE).}
}
\value{
Named list of similarity scores.
}
\description{
Uses the NBLAST algorithm that compares the morphology of two neurons. For
more control over the parameters of the algorithm, see the arguments of
\code{\link{NeuriteBlast}}.
}
\details{
when \code{smat=NULL} options("nat.nblast.defaultsmat") will be
  checked and if NULL, then \code{smat.fcwb} or \code{smat_alpha.fcwb} will
  be used depending on the value of \code{UseAlpha}.
}
\examples{
data(kcs20,package='nat')
nblast(kcs20[[1]],kcs20)
}
\seealso{
\code{\link{nat-package}}
}

