% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ndclust}
\alias{ndclust}
\title{Cluster NBLAST scores using density clustering algorithm}
\usage{
ndclust(score_matrix, ...)
}
\arguments{
\item{score_matrix}{a raw NBLAST score matrix.}

\item{...}{extra arguments to pass to
  \code{\link[densityClust]{densityClust}}.}
}
\value{
An \code{ndclust} object (basically a \code{densityCluster} object,
  see \link[densityClust]{densityClust}) with clusters assigned to all
  neurons. See \code{\link[densityClust]{findClusters}}.
}
\description{
Cluster NBLAST scores using density clustering algorithm
}

