% Generated by roxygen2 (4.0.1): do not edit by hand
\name{NeuriteBlast}
\alias{NeuriteBlast}
\title{Produce similarity score for neuron morphologies}
\usage{
NeuriteBlast(query, target, targetBinds = NULL, normalised = FALSE, ...)
}
\arguments{
\item{query}{either a single query neuron or a \code{\link[nat]{neuronlist}}}

\item{target}{a \code{neuronlist} to compare neuron against.}

\item{targetBinds}{numeric indices or names with which to subset
\code{target}.}

\item{normalised}{whether to divide scores by the self-match score of the
query}

\item{...}{extra arguments to pass to the distance function.}
}
\value{
Named list of similarity scores.
}
\description{
A low-level version of the NBLAST algorithm that compares the morphology of a
neuron with those of a list of other neurons. For most use cases, one would
probably wish to use \code{\link{nblast}} instead.
}
\seealso{
\code{\link{WeightedNNBasedLinesetMatching}}
}

