% Generated by roxygen2 (4.0.1): do not edit by hand
\name{neuron_pairs}
\alias{neuron_pairs}
\title{Utility function to generate all or random pairs of neurons}
\usage{
neuron_pairs(query, target, n = NA, includeSelf = FALSE)
}
\arguments{
\item{query,target}{either neuronlists or character vectors of names. If
target is missing, query will be used as both query and target.}

\item{n}{number of random pairs to draw. When NA, the default, uses
\code{expand.grid} to draw all pairs.}
}
\value{
a data.frame with two character vector columns, query and target.
}
\description{
Utility function to generate all or random pairs of neurons
}
\examples{
neuron_pairs(nat::kcs20, n=20)
}
\seealso{
\code{\link{calc_score_matrix}, \link{expand.grid}}
}

