% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot3d.ndclust}
\alias{plot3d.ndclust}
\title{Plot neurons in 3D based on density clustering}
\usage{
\method{plot3d}{ndclust}(x, groups = NULL, col = rainbow, ...)
}
\arguments{
\item{x}{an \code{\link{ndclust}} object detailing the clustering of neurons.}

\item{groups}{numeric vector of groups to plot.}

\item{col}{colours for groups (directly specified or a function).}

\item{...}{additional arguments for \code{plot3d}.}
}
\value{
A list of rgl IDs for plotted objects (see \code{\link[rgl]{plot3d}}).
}
\description{
Plot neurons in 3D based on density clustering
}

