% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot3d.sclust}
\alias{plot3d.sclust}
\title{Plot the results of a spectral clustering}
\usage{
\method{plot3d}{sclust}(x, col = rainbow, ...)
}
\arguments{
\item{x}{an \code{\link{sclust}} object.}

\item{col}{colours for groups (directly specified or a function).}

\item{...}{additional arguments for \code{plot3d}}
}
\value{
A list of rgl IDs for plotted objects (see \code{\link[rgl]{plot3d}}).
}
\description{
Plot the results of a spectral clustering
}

