% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_hits_from_matrix}
\alias{plot_hits_from_matrix}
\title{Display NBLAST hits for neurons in 3D, given scores}
\usage{
plot_hits_from_matrix(nblast_results, neurons, threshold = NULL,
  number = NULL, col = heat.colors, ...)
}
\arguments{
\item{nblast_results}{a matrix of NBLAST scores.}

\item{neurons}{a neuronlist of neurons to plot from.}

\item{threshold}{an optional number which, if given, prevents scores below
this number from being displayed.}

\item{number}{an optional number of neurons to plot.}

\item{col}{the colour palette with which to plot the neurons.}

\item{...}{extra parameters to pass to plot3d.}
}
\value{
\code{NULL}.
}
\description{
Display NBLAST hits for neurons in 3D, given scores
}

