% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_hits_from_scores}
\alias{plot_hits_from_scores}
\title{Display NBLAST hits for a neuron in 3D, given scores}
\usage{
plot_hits_from_scores(nblast_results, neurons, threshold = NULL,
  col = heat.colors, ...)
}
\arguments{
\item{nblast_results}{a vector of NBLAST scores.}

\item{neurons}{a neuronlist of neurons to plot from.}

\item{threshold}{an optional number which, if given, prevents scores below
this number from being displayed.}

\item{col}{the colour palette with which to plot the neurons.}

\item{...}{extra parameters to pass to plot3d.}
}
\value{
A list of rgl IDs (invisibly).
}
\description{
Display NBLAST hits for a neuron in 3D, given scores
}

