% Generated by roxygen2 (4.0.1): do not edit by hand
\name{random_pairs}
\alias{random_pairs}
\title{Utility function for generating random pairs of neurons}
\usage{
random_pairs(x, n)
}
\arguments{
\item{x}{neuronlist or character vector of names}

\item{n}{number of random pairs to draw}
}
\value{
a data.frame with two character vector columns, query and target.
}
\description{
Utility function for generating random pairs of neurons
}
\examples{
random_pairs(nat::kcs20, n=20)
}
\seealso{
calc_score_matrix
}

