% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sclust}
\alias{sclust}
\title{Cluster NBLAST scores using spectral clustering}
\usage{
sclust(x, num_clusters, ...)
}
\arguments{
\item{x}{the NBLAST score matrix to cluster.}

\item{num_clusters}{the number of desired clusters.}

\item{...}{extra arguments to pass to \code{\link{kmeans}}.}
}
\value{
A object of class \code{sclust} containing the final k-mean
  clustering object, along with a vector of labels.
}
\description{
Cluster NBLAST scores using spectral clustering
}

