% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scree_plot}
\alias{scree_plot}
\title{Plot a scree plot for NBLAST scores}
\usage{
scree_plot(scores, max_k, num_runs = 1, perm = NULL, ...)
}
\arguments{
\item{scores}{a matrix of NBLAST scores.}

\item{max_k}{the maximum number of clusters to consider.}

\item{num_runs}{the number of times to repeat each k-means clustering for a
given number of clusters.}

\item{perm}{either \code{NULL}, indicating no permutations should be
considered, or the number of times the data should be permuted.}

\item{...}{extra arguments to pass to \code{\link{plot}}.}
}
\value{
(Invisibly) A list of within-cluster sum-of-squares for each number of clusters
  tested, for both real and (if \code{perm} is not \code{NULL}) permuted
  data.
}
\description{
This plots the number of considered clusters for a k-means clustering
(x-axis) against the within-cluster sum-of-squares (y-axis). As the number of
considered clusters increases the within-cluster sum-of-squares drops,
typically fairly steeply at first and then followed by an elbow to a more
gradual, almost level descent. Cattell's scree test states that the maximum
number of 'proper' clusters is the one at which this elbow occurs.
}
\details{
If a number is passed in as \code{perm}, the within-cluster sum-of-squares
for random permutations of data will also be displayed. If a dataset has
'prominent' clusters then one would expect that the within-cluster
sum-of-squares for the real data would decrease faster than those of the
permuted versions. A message indicating the maximum difference in
within-cluster sum-of-squares for original and permuted data is also
displayed.
}

