% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.r
\name{subset.hclust}
\alias{subset.hclust}
\title{Return the labels of items in 1 or more groups cut from hclust object}
\usage{
\method{subset}{hclust}(x, k = NULL, h = NULL, groups = NULL, ...)
}
\arguments{
\item{x}{tree like object}

\item{k}{an integer scalar with the desired number of groups}

\item{h}{numeric scalar with height where the tree should be cut}

\item{groups}{a vector of which groups to inspect.}

\item{...}{Additional parameters passed to methods}
}
\value{
A character vector of labels of selected items
}
\description{
Return the labels of items in 1 or more groups cut from hclust object
}
\details{
Only one of \code{h} and \code{k} should be supplied.
}

