% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create_smat}
\alias{create_smat}
\title{Create a scoring matrix given matching and non-matching sets of neurons}
\usage{
create_smat(matching_neurons, nonmatching_neurons, ignoreSelf = TRUE,
  distbreaks, dotprodbreaks = seq(0, 1, by = 0.1), logbase = 2,
  fudgefac = 1e-06, ...)
}
\arguments{
\item{matching_neurons}{a \code{\link[nat]{neuronlist}} of matching neurons.}

\item{nonmatching_neurons}{a \code{\link[nat]{neuronlist}} of non-matching
neurons.}

\item{ignoreSelf}{a Boolean indicating whether to ignore comparisons of a
neuron against itself (default \code{TRUE}).}

\item{distbreaks}{a vector specifying the breaks for distances in the
probability matrix.}

\item{dotprodbreaks}{a vector specifying the breaks for dot products in the
probability matrix.}

\item{logbase}{the base to which the logarithm should be taken to produce the
final scores.}

\item{fudgefac}{an arbitrary, small 'fudge factor' to prevent division by
zero.}

\item{...}{extra arguments to pass to \code{\link{NeuriteBlast}}.}
}
\value{
A matrix with columns as specified by \code{dotprodbreaks} and rows
  as specified by \code{distbreaks}, containing log odd scores for neuron
  segments with the given distance and dot product.
}
\description{
Create a scoring matrix given matching and non-matching sets of neurons
}

