% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{nat.nblast-package}
\alias{nat.nblast}
\alias{nat.nblast-package}
\title{Neuron similarity, search and clustering tools}
\description{
Neuron similarity, search and clustering tools
}
\references{
Costa, M., Ostrovsky, A.D., Manton, J.D., Prohaska, S., and Jefferis, G.S.X.E. (2014).
NBLAST: Rapid, sensitive comparison of neuronal structure and construction of neuron family databases.
Biorxiv preprint. \href{http://dx.doi.org/10.1101/006346}{doi: 10.1101/006346}
}
\keyword{package}

