% Generated by roxygen2 (4.0.1): do not edit by hand
\name{nblast}
\alias{nblast}
\title{Produce similarity score for neuron morphologies}
\usage{
nblast(query, target, smat = get(getOption("nat.nblast.defaultsmat")),
  version = c(2, 1), UseAlpha = NULL)
}
\arguments{
\item{query}{the query neuron.}

\item{target}{a \code{\link[nat]{neuronlist}} to compare neuron against.
Defaults to \code{options("nat.default.neuronlist")}.}

\item{smat}{the score matrix to use.}

\item{version}{the version of the algorithm to use (the default, 2, is the
latest).}

\item{UseAlpha}{whether to consider local directions in the similarity
calculation. The default value, \code{NULL} ensures that the canonical
setting for each version of NBLAST is used.}
}
\value{
Named list of similarity scores.
}
\description{
Uses the NBLAST algorithm that compares the morphology of two neurons. For
more control over the parameters of the algorithm, see the arguments of
\code{\link{NeuriteBlast}}.
}
\examples{
data(kcs20,package='nat')
nblast(kcs20[[1]],kcs20)
}

