% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot_hits}
\alias{plot_hits}
\title{Calculate and display NBLAST hits for a neuron in 3D}
\usage{
plot_hits(query, target, threshold = NULL, number = NULL,
  col = heat.colors, ...)
}
\arguments{
\item{query}{the query neuron.}

\item{target}{a \code{neuronlist} to compare against.}

\item{threshold}{an optional number which, if given, prevents scores below
this number from being displayed.}

\item{number}{an optional number of neurons to plot.}

\item{col}{the colour palette with which to plot the neurons.}

\item{...}{extra parameters to pass to plot3d.}
}
\value{
A list of rgl IDs (invisibly).
}
\description{
Calculate and display NBLAST hits for a neuron in 3D
}

