% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sparse_score_mat}
\alias{sparse_score_mat}
\title{Convert a subset of a score matrix to a sparse representation}
\usage{
sparse_score_mat(neuron_names, dense_matrix)
}
\arguments{
\item{neuron_names}{a character vector of neuron names to save scores for.}

\item{dense_matrix}{the original, dense version of the full score matrix.}
}
\value{
A spare matrix, in compressed, column-oriented form, as an R object
  inheriting from both \code{\link[Matrix]{CsparseMatrix-class}} and
  \code{\link[Matrix]{generalMatrix-class}}.
}
\description{
This can be useful for storing raw forwards and reverse NBLAST scores for a
set of neurons without having to store all the uncomputed elements in the
full score matrix.
}

