% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sub_dist_mat}
\alias{sub_dist_mat}
\title{Return a subset of a distance matrix stored in a file-backed matrix}
\usage{
sub_dist_mat(neuron_names, scoremat = NULL, form = c("matrix", "dist"),
  maxneurons = NA)
}
\arguments{
\item{form}{the type of object to return.}

\item{maxneurons}{set this to a sensible value to avoid loading huge (order
N^2) distances directly into memory.}

\item{neuron_names}{character vector of neuron identifiers.}

\item{scoremat}{score matrix to use (see \code{sub_score_mat} for details of
default).}
}
\value{
return An object of class matrix or dist (as determined by the form
  argument), corresponding to a subset of the distance matrix
}
\description{
Return a subset of a distance matrix stored in a file-backed matrix
}
\seealso{
Other scoremats: \code{\link{hclustn}}
}

